Simplify configuration for fixture files
By using a karma proxy we can use the built-in fixtures path of jasmine-jquery without having to customise the path in each spec file and let is simply serve files from src/angularjs. To simplify even further, we add a default file pattern to the 'files' option to serve fixture files.