Update readme on new fixture configuration
Hippo-build will provide out-of-the-box fixture support as long as there is no need to customise the Karma 'files' and/or 'proxies' options. Should this be needed (which is pretty likely) it should be clear from the documentation on how to do that without breaking fixture support (if neeed at all).