FEBUILD-13 remove babel-eslint and es6-module-loader
According to the babel-eslint docs, "You don't need to use babel-eslint
if you are only using ES6/ES2015". This also fixes the wrong dependency
of babel-eslint on estraverse (eslint #5476).

The usage of ES6 module loader broke the build since it was no longer
pulled in by SystemJs. Turns out we actually don't need it.