FEBUILD-49 Use CleanWebpackPlugin to remove dist folder
Gulp tasks can return a promise to handle signaling done, which is nice, but results in ugly/wrong exit status&output
when a task is executed in series and throws an error.

This happens to all the build tasks, which are executed in series with the gulp 'clean' task, which simply removes the
dist folder. Fortunately Webpack has a CleanWebpackPlugin.